Fix default setting bug
